HitFilterAlg made optional in GausHitFinder
While filtering was optional, it was necessary to always specify a configuration for the hit filter, regardless it was then used or not.
Now the hit filter algorithm is not instantiated unless filtering is required.
If filtering is not required (that used to be and still is the default), no further configuration is required.