FuzzyClusterMerger now excludes hits with negative charge.
Hit with negative charge (recob::Hit::Integral()) are not supposed to
appear. Some high concentration of them in a cluster can make its total
charge become negative and weird things happen.
This patch implements an arbitrary choice of coping with the faulty
input by ignoring all the hits with negative charge. If a cluster is
completely depleted, that cluster will also be ignored.
Note that this is basically a patch to survive upstream negligence.