Prepare DisambigCheater for geometry change.
DisambigCheater was not prepared to exceptions caused by GeometryCore::NearestWireID() on points outside the target plane.
DUNE, the only user so far, had a special implementation which would
handle the error and provide directly the "best" result.
The generic implementation actually stores that very same "best" result
in the exception it throws, and this patch is going to use that
information when needed.
This is a completely backward compatible change.