Add switch to enable SpacePointSolver to consider triplets where one of the hits is actually a bad channel.
Currently this bad channel is constrained to be in the V view. And it's off by default. The reason not to implement the U view (requires a bunch of restructuring to do it neatly), or to enable by default, is that it seems to generate quite a lot of ghost space points.